+++ src/main/java/com/exaroton/api/request/server/files/GetConfigOptionsRequest.java
@@ -1,25 +1,26 @@
 package com.exaroton.api.request.server.files;
 
-import com.exaroton.api.APIRequest;
 import com.exaroton.api.APIResponse;
 import com.exaroton.api.ExarotonClient;
+import com.exaroton.api.server.config.ConfigOption;
 import com.google.gson.reflect.TypeToken;
 
 import java.lang.reflect.Type;
-import java.util.HashMap;
 
-public class FileDataRequest extends FileRequest<Object> {
-    public FileDataRequest(ExarotonClient client, String serverId, String path) {
+public class GetConfigOptionsRequest extends FileRequest<ConfigOption[]> {
+
+    public GetConfigOptionsRequest(ExarotonClient client, String serverId, String path) {
         super(client, serverId, path);
     }
 
+
     @Override
     protected String getEndpoint() {
-        return "servers/{server}/files/data/{path}";
+        return "servers/{server}/files/config/{path}";
     }
 
     @Override
     protected Type getType() {
-        return new TypeToken<APIResponse<Object>>(){}.getType();
+        return new TypeToken<APIResponse<ConfigOption[]>>(){}.getType();
     }
 }