+++ src/main/java/com/exaroton/api/server/Server.java
@@ -382,7 +382,7 @@ public class Server {
     public void subscribe() {
         String protocol = this.client.getProtocol().equals("https") ? "wss" : "ws";
         String uri = protocol + "://" + this.client.getHost() + this.client.getBasePath() + "servers/" + this.id + "/websocket";
-        this.webSocket = new WebSocketManager(uri, this.client.getApiToken(), this);
+        this.webSocket = new WebSocketManager(client, uri, this.client.getApiToken(), this);
     }
 
     /**