+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- Add option to use own credits to Server.start()
-- Use JSON body instead of form data
-- Update dependencies
+- Add endpoints for the server config